requirements.txt
azure-functions
flask
azure-storage-blob
